fix: 별점 타입 Integer

+++ src/main/java/com/ureca/gate/review/infrastructure/jpaadapter/entity/ReviewEntity.java
@@ -19,7 +19,7 @@ public class ReviewEntity extends BaseTimeEntity {
 
   private Long placeId;
 
-  private Double starRate;
+  private Integer starRate;
 
   private String size;
